+++ ChangeLog.md
@@ -5,6 +5,7 @@
 ## 1.6
 
  * Add mtev_set_app_name and mtev_get_app_name for conf file root discovery
+ * Add foreground SIGINT, SIGQUIT, and SIGTERM signals to call plain-old exit()
 
 ### 1.6.16